// src/lib/cmf/types.d.ts
export interface TextEvent {
    /**
     * Timestamp in milliseconds when this event occurs.
     */
    t: number;
    type: "text";
    /**
     * The FEN string of the position this event relates to.
     */
    fen: string;
    /**
     * The textual content to display.
     */
    text: string;
}